refactor: split comment_verdict into protocol-type, entrypoint and pre-launch siblings
comment_verdict.py had grown to 1551 lines and failed the first-party
line-limit guardrail, and trimming it to just under 1500 leaves no room for
the next review round. Move three self-contained blocks into sibling modules,
each re-exported (``X as X``) so ``comment_verdict`` stays the import and
monkeypatch surface for ``comments`` and the tests:

- comment_verdict_protocol_types: reason codes, the verdict Literals and the
  ``VerdictResult`` / ``MonitorVerdictResult`` / error types plus the two
  correction prompt constants. No I/O and no sibling imports.
- comment_verdict_entrypoint: ``_invoke_cli_for_verdict`` and
  ``_invoke_cli_for_verdict_result``, which re-arm and staleness-check the
  #932 anchor around the protocol. It resolves ``_run_item_verdict_protocol``
  through ``comment_verdict`` at call time, so the existing monkeypatch seam
  and the ``comments`` forwarding shim keep working.
- comment_verdict_prelaunch: the once-per-item anchor resolution and
  pre-launch repairs, returning the resolved anchors (including the mirror
  path and the HEAD reader, so no later probe re-derives them) as a frozen
  ``ItemProtocolAnchors``. Ownership repair, mirror lookup, hooks repair and
  the git-config snapshot are resolved through ``comment_verdict`` because
  tests patch them there.

``mirror_path_for_worktree`` / ``repair_agent_runtime_ownership`` become
explicit re-exports, which retires the two now-unused ``type: ignore``
comments in the ``comments`` compatibility seam. Behaviour is unchanged:
comment_verdict.py is 1551 -> 1285 lines and every new module is well under
the limit.
