+++ src/awf/runtime/pr_monitor_runner/comments.py
@@ -269,8 +269,8 @@ def _agent_failed_result(exc: AgentVerdictExecutionError) -> MonitorVerdictResul
 
 def _sync_comment_verdict_dependencies() -> None:
     """Keep legacy comment-module monkeypatch seams for verdict invocation tests."""
-    _comment_verdict.mirror_path_for_worktree = mirror_path_for_worktree  # type: ignore[attr-defined]
-    _comment_verdict.repair_agent_runtime_ownership = repair_agent_runtime_ownership  # type: ignore[attr-defined]
+    _comment_verdict.mirror_path_for_worktree = mirror_path_for_worktree
+    _comment_verdict.repair_agent_runtime_ownership = repair_agent_runtime_ownership
     _comment_verdict.repair_mirror_hooks_path = repair_mirror_hooks_path
     _comment_verdict.mirror_hooks_repair_failure_details = mirror_hooks_repair_failure_details  # type: ignore[attr-defined]
